refactor: extract merge functions into coverageDataMerger module
Moves mergeDeployCoverageData and mergeTestCoverageData out of
coverageTransformer.ts to reduce its total complexity score.
